Move the logic of Ax plugin in a separate file
Now Ax is imported lazily
